backend/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
openai==1.63.2
python-dotenv==1.0.1
pymongo==4.11.1
pydantic[email]==2.10.6
PyMuPDF==1.25.3
tools==0.1.9
python-docx==1.1.2
python-multipart==0.0.20
openai==1.63.2
pinecone-client==5.0.1
langchain-openai==0.3.8